Add platform-specific native library loading and enhance build scripts
- Implemented platform detection in build-native.sh for Linux, macOS, and Windows.
- Updated JNI header generation to include multiple Java classes.
- Created NativeLibraryLoader for dynamic loading of native libraries based on the platform.
- Added new enums for HighsStatus, ModelStatus, and VarType to improve code organization.
- Enhanced Maven build configuration to support packaging and dependency management.
- Updated README with installation and usage instructions.
