+++ pom.xml
@@ -64,7 +64,13 @@
                         </goals>
                         <configuration>
                             <executable>javac</executable>
-                            <commandlineArgs>-h ${project.basedir}/src/main/native/headers src/main/java/nl/jessenagel/jhighs/HiGHS.java</commandlineArgs>
+                            <arguments>
+                                <argument>-h</argument>
+                                <argument>${project.build.directory}/generated-sources/jni</argument>
+                                <argument>-classpath</argument>
+                                <argument>${project.build.outputDirectory}</argument>
+                                <argument>src/main/java/nl/jessenagel/jhighs/HiGHS.java</argument>
+                            </arguments>
                         </configuration>
                     </execution>
                     <execution>
@@ -81,18 +87,17 @@
                 </executions>
             </plugin>
 
-            <!-- Surefire for running tests -->
+            <!-- Enhanced Surefire for running tests -->
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>3.2.5</version>
                 <configuration>
-                        <argLine>-Djava.library.path=${project.basedir}/src/main/resources/natives/linux-x86_64</argLine>
-                        <environmentVariables>
-                            <LD_LIBRARY_PATH>${project.basedir}/src/main/resources/natives/linux-x86_64</LD_LIBRARY_PATH>
-                        </environmentVariables>
+                    <argLine>-Djava.library.path=${project.basedir}/src/main/resources/natives/linux-x86_64</argLine>
+                    <environmentVariables>
+                        <LD_LIBRARY_PATH>${project.basedir}/src/main/resources/natives/linux-x86_64</LD_LIBRARY_PATH>
+                    </environmentVariables>
                 </configuration>
-
             </plugin>
 
             <!-- Include native libraries in JAR -->
@@ -104,7 +109,62 @@
                     <includeEmptyDirs>true</includeEmptyDirs>
                 </configuration>
             </plugin>
+
+            <!-- Create fat JAR with all dependencies -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-shade-plugin</artifactId>
+                <version>3.5.0</version>
+                <executions>
+                    <execution>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>shade</goal>
+                        </goals>
+                        <configuration>
+                            <createDependencyReducedPom>false</createDependencyReducedPom>
+                            <filters>
+                                <filter>
+                                    <artifact>*:*</artifact>
+                                    <excludes>
+                                        <exclude>META-INF/*.SF</exclude>
+                                        <exclude>META-INF/*.DSA</exclude>
+                                        <exclude>META-INF/*.RSA</exclude>
+                                    </excludes>
+                                </filter>
+                            </filters>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+
+            <!-- Alternative: Assembly plugin for more control -->
+            <!--
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-assembly-plugin</artifactId>
+                <version>3.6.0</version>
+                <configuration>
+                    <descriptorRefs>
+                        <descriptorRef>jar-with-dependencies</descriptorRef>
+                    </descriptorRefs>
+                    <archive>
+                        <manifest>
+                            <mainClass>your.main.Class</mainClass>
+                        </manifest>
+                    </archive>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>make-assembly</id>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>single</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            -->
         </plugins>
     </build>
-
 </project>
